+++ app/modules/legal/__init__.py
@@ -1,0 +1,1 @@
+"""Play Store / privacy compliance public APIs."""